Add arch scripts and update other scripts

+++ build/debian/overlay/penv/cleanup.d/01-debian.sh
@@ -1,10 +1,6 @@
 #!/bin/sh
 set -e
 
-if [ "$PENV_ENV_MODE" != "build" ] && [ "$PENV_ENV_MODE" != "mod" ] && [ -z "$PENV_SIGNAL_CLEANUP" ]; then
-    exit 0
-fi
-
 VERBOSE=${PENV_CONFIG_VERBOSE:-0}
 [ "${1:-}" = "-v" ] && VERBOSE=1
 
@@ -15,13 +11,19 @@ else
     RM_FLAGS="-rf"
 fi
 
+# Remove apt cache
+rm $RM_FLAGS /var/cache/apt || true
+
+if [ "$PENV_ENV_MODE" != "build" ] && [ "$PENV_ENV_MODE" != "mod" ] && [ -z "$PENV_SIGNAL_CLEANUP" ]; then
+    exit 0
+fi
+
 if [ $VERBOSE -eq 1 ]; then
     echo "Cleaning Debian-specific files..."
 fi
 
-# Remove apt lists and cached debs
+# Remove apt lists
 rm $RM_FLAGS /var/lib/apt/lists/* || true
-rm $RM_FLAGS /var/cache/apt || true
 
 # remove apt logs
 for log in /var/log/apt/history.log /var/log/apt/term.log; do
